% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_eval.R
\name{augment_cdf}
\alias{augment_cdf}
\title{Augment a CDF to full domain and range}
\usage{
augment_cdf(dat, xrange, yrange = 0:1)
}
\arguments{
\item{dat}{Data frame. The CDF information from \code{dprocsim::run_shiny()}. Should have columns named \code{x}, \code{y}, \code{bound}, \code{method}.}

\item{xrange}{Numeric. Vector of length 2 giving the minimum and maximum of the function's domain.}

\item{yrange}{Numeric. Vector of length 2 giving the minimum and maximum of the fundtion's range. Should be \code{0:1} for CDF (default).}
}
\description{
Augment a CDF to full domain and range
}
