% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cdfs.R
\name{check_cdf_bounds}
\alias{check_cdf_bounds}
\alias{check_whole_cdf}
\alias{check_cdf_many}
\alias{check_cdf_counts}
\title{Check CDF values fall in between bounds}
\usage{
check_cdf_bounds(dpp_draw, ubf, lbf, x)

check_whole_cdf(dpp_draw, ubf, lbf, xgrid)

check_cdf_many(dpp_draws, ubf, lbf, xgrid)

check_cdf_counts(many_draws_checked, N)
}
\arguments{
\item{dpp_draw}{Data frame. A single draw of a CDF from a Dirichlet Process. Output of the \code{dp*_sim*()} functions. Should have columns names \code{locations} and \code{weCDF}.}

\item{ubf}{Function. The upper bound function created with the clicked data points. Output of either \code{spline_funs()} or \code{linear_funs()}.}

\item{lbf}{Function. The lower bound function created with the clicked data points. Output of either \code{spline_funs()} or \code{linear_funs()}.}

\item{x}{Numeric. A single x-value to check in the domain of the CDF.}

\item{xgrid}{Numeric. A vector of x values in the domain of the CDF to check. It is recommended you don't put the bounds in.}

\item{dpp_draws}{A tibble. The output from \code{dpprior_sim2()} (or \code{dppost_sim2()}).}

\item{many_draws_checked}{A tibble. The output from \code{check_cdf_many()}}

\item{N}{An integer. How many repeated draws are in \code{many_draws_checked}? Same \code{N} as that of \code{dp*_sim2()}.}
}
\description{
Check CDF values fall in between bounds

Check CDF over a grid of x values

Check many CDFs over a grid of x values

Determine regions of random CDFs.
}
\section{Functions}{
\itemize{
\item \code{check_whole_cdf}: 

\item \code{check_cdf_many}: 

\item \code{check_cdf_counts}: 
}}

