% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cdfs.R
\name{check_cdf_dist}
\alias{check_cdf_dist}
\title{Compute distance from given bounds}
\usage{
check_cdf_dist(cdf_regions, au, bl, bw)
}
\arguments{
\item{cdf_regions}{A tibble. Output from \code{check_cdf_counts()}}

\item{au}{Numeric. A number between 0-1. What proportion of simulated CDFs should lie above the upper bound?}

\item{bl}{Numeric. A number between 0-1. What proportion of simulated CDFs should lie below the lower bound?}

\item{bw}{Numeric. A number between 0-1. What proportion of simulated CDFs should between the upper and lower bounds?}
}
\description{
Compute distance from given bounds
}
