% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_proc_sim.R
\name{dppost_sim}
\alias{dppost_sim}
\title{Simulate from a DP Posterior}
\usage{
dppost_sim(M, F0, sticks, dat, ...)
}
\arguments{
\item{M}{An integer. The precision parameter in a Dirichlet Process Prior.}

\item{F0}{A character. Name of the function for simulating from the desired centering base measure.}

\item{sticks}{An integer. The number of sticks to break when simulating from DP(M, F0) using the stick breaking construction.}

\item{dat}{A vector. Observed data.}

\item{...}{Named arguments to pass to to the centering function.}
}
\description{
Simulate from a DP Posterior
}
\examples{
dat <- runif(25, 0, 1)
beta2 <- dppost_sim(dat, M = 4, F0 = "rbeta", sticks = 100, shape1 = 4, shape2 = 1)
dat <- runif(25, .75, 1.25)
norm2 <- dppost_sim(dat, M = 1, F0 = "rnorm", sticks = 100, mean = 1, sd = .01)
}
