% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_proc_sim.R
\name{dppost_sim2}
\alias{dppost_sim2}
\title{Simulate many Distributions from DP Posteriors}
\usage{
dppost_sim2(M, F0, sticks, dat, N, ...)
}
\arguments{
\item{M}{An integer. The precision parameter in a Dirichlet Process Prior.}

\item{F0}{A character. Name of the function for simulating from the desired centering base measure.}

\item{sticks}{An integer. The number of sticks to break when simulating from DP(M, F0) using the stick breaking construction.}

\item{dat}{A vector. Observed data.}

\item{N}{An integer. The number of distributions to simulate for each value of \code{M}.}

\item{...}{Named arguments to pass to to the centering function.}
}
\description{
Simulate many Distributions from DP Posteriors
}
