% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_proc_sim.R
\name{dpprior_sim2}
\alias{dpprior_sim2}
\title{Simulate many Distributions from DP Priors}
\usage{
dpprior_sim2(M, F0, sticks, N, ...)
}
\arguments{
\item{M}{Integer. Value of the scaling parameter in the DP prior. May be a vector.}

\item{F0}{A character. Function for the centering distribution in the DP prior.}

\item{sticks}{An integer. The number of sticks to break when simulating from \eqn{DP(M, F_0)} using the stick breaking construction.}

\item{N}{An integer. The number of distributions to simulate for each value of \code{M}.}

\item{...}{Named arguments to pass to to the centering function.}
}
\description{
Obtain repeated draws from a (set of) given DP priors.
}
\examples{
M <- 1:2
F0 <- "rbeta"

dpprior_sim2(M = M, F0 = rbeta, sticks = 100, N = 10, shape1 = 4, shape2 = 1)
}
