% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprocsim.R
\docType{package}
\name{dprocsim}
\alias{dprocsim}
\alias{dprocsim-package}
\title{\code{dprocsim} package}
\description{
Simulate from Dirichlet Process priors and posteriors, and estimate posterior densities.
}
