% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_eval.R
\name{eval_dens}
\alias{eval_dens}
\title{Density Estimation for a DP Posterior}
\usage{
eval_dens(x, epsilon, sims, ...)
}
\arguments{
\item{x}{A numeric. Value in the domain at which to evaluate the posterior.}

\item{epsilon}{A numeric. Small value around which to estimate the density.}

\item{sims}{A tibble. Simulations from \code{\link[=dppost_sim2]{dppost_sim2()}} (preferred) or \code{\link[=dppost_sim]{dppost_sim()}}.}

\item{...}{Named arguments to pass to to the centering function.}
}
\description{
Compute density esimates for a value by computing difference in distribution values for a small interval around value of interest.
}
