% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_proc_sim.R
\name{F_0}
\alias{F_0}
\title{Simulate from Prior Centering Measure}
\usage{
F_0(sticks, F0, ...)
}
\arguments{
\item{sticks}{An integer. The number of sticks to break when simulating from \eqn{DP(M, F_0)} using the stick breaking construction.}

\item{F0}{A character. Name of the function for simulating from the desired centering base measure.}

\item{...}{Named arguments to pass to to the centering function.}
}
\description{
Prior form of the centering measure.
}
