% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_proc_sim.R
\name{F_0_prime}
\alias{F_0_prime}
\title{Simulate from Posterior Centering Measure}
\usage{
F_0_prime(M, F0, sticks, dat, ...)
}
\arguments{
\item{M}{An integer. The precision parameter in a Dirichlet Process Prior.}

\item{F0}{A character. Name of the function for simulating from the desired centering base measure.}

\item{sticks}{An integer. The number of sticks to break when simulating from \eqn{DP(M, F_0)} using the stick breaking construction.}

\item{dat}{A vector. Observed data.}

\item{...}{Named arguments to pass to to the centering function.}
}
\description{
Posterior (conjugate) form of the centering measure.
}
