% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_eval.R
\name{in_range}
\alias{in_range}
\title{Check if a Value Falls in Interval}
\usage{
in_range(location, fixed_value, epsilon)
}
\arguments{
\item{location}{A numeric. Value to check.}

\item{fixed_value}{A numeric. Center of the interval.}

\item{epsilon}{A numeric. Half of interval width.}
}
\description{
Use this function if you want to perform interval density esimates by hand.
}
