% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Run Built-In Shiny App(s)}
\usage{
run_shiny(appname = "cdf-picker")
}
\arguments{
\item{appname}{Character. Name of the app to run. There is only one option: \code{"cdf-picker"}}
}
\description{
This app allows you to draw upper and lowers bounds for a CDF. When simulating from Dirichlet Process Priors with centering distribution as the mean of the upper and lower bounds drawns, we expect that some CDFs simulated will fall outside of the given bounds. Be thinking about what proportion of simulated CDFs you want to fall outside of
}
\examples{
\dontrun{
run_shiny()
}

}
