% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawn-cdf-sim.R
\name{spline_funs}
\alias{spline_funs}
\alias{linear_funs}
\alias{mean_funs}
\title{Helpful closures}
\usage{
spline_funs(clicks, xrange, spar = 0.5)

linear_funs(clicks, xrange)

mean_funs(upper, lower, meth)
}
\arguments{
\item{clicks}{Clicked data}

\item{xrange}{Vector of length 2 giving the minimum and maximum domain values for the spline function.}

\item{spar}{Numeric. Value of the smoothing parameter (between 0-1) to be passed to \code{smooth.spline()}.}

\item{upper}{Data frame. The upper bound of points for the cdf}

\item{lower}{Data frame. The lower bound of points for the cdf}

\item{meth}{Character. Approximation method. either \code{"splines"} for \code{smooth.spline()} or \code{"linear"} for \code{approxfun()}}
}
\description{
Create linear approximation or spline approximation functions from a given set of points.
}
\section{Functions}{
\itemize{
\item \code{linear_funs}: Linear function closure

\item \code{mean_funs}: Mean function closure
}}

\examples{
\dontrun{
splinefuns(clicks)(seq(xmin, xmax, length.out = 100))
}

}
